[Refactor] 불필요한 추상클래스 AbstractKnockedPinCounts 제거
