+++ src/main/java/bowling/domain/PinCounts.java
@@ -1,19 +1,30 @@
 package bowling.domain;
 
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class PinCounts extends AbstractKnockedPinCounts {
-    private static final int MAX_SIZE = 2;
+public class PinCounts implements KnockedPinCounts {
+    public static final String WRONG_BOWL_COUNT_MESSAGE = "잘못된 투구 수입니다.";
+
+    private final List<KnockedPinCount> values;
+
+    public PinCounts(List<KnockedPinCount> knockedPinCounts, int maxSize) {
+        this.values = knockedPinCounts;
+
+        if (knockedPinCounts.size() > maxSize) {
+            throw new IllegalArgumentException(WRONG_BOWL_COUNT_MESSAGE);
+        }
+    }
 
     public PinCounts(KnockedPinCount... knockedPinCounts) {
-        this(Arrays.stream(knockedPinCounts)
-                .collect(Collectors.toList()));
+        this.values = Arrays.stream(knockedPinCounts)
+                .collect(Collectors.toList());
     }
 
     public PinCounts(List<KnockedPinCount> knockedPinCounts) {
-        super(knockedPinCounts, MAX_SIZE);
+        this(knockedPinCounts, MAX_SIZE);
     }
 
     @Override
@@ -21,4 +32,10 @@ public class PinCounts extends AbstractKnockedPinCounts {
         values.add(new KnockedPinCount(knockedOutCount));
         return this;
     }
+
+    @Override
+    public List<KnockedPinCount> getValues() {
+        return Collections.unmodifiableList(values);
+    }
+    private static final int MAX_SIZE = 2;
 }
